Login added and no result ui moded

+++ app/src/main/java/com/ass/madhwavahini/ui/presentation/MainViewModel.kt
@@ -1,7 +1,9 @@
 package com.ass.madhwavahini.ui.presentation
 
 import android.app.Application
+import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.setValue
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.ass.madhwavahini.data.local.UserDataStore
@@ -9,13 +11,13 @@ import com.ass.madhwavahini.data.local.dao.FilesDao
 import com.ass.madhwavahini.domain.modals.Payment
 import com.ass.madhwavahini.domain.modals.User
 import com.ass.madhwavahini.domain.repository.PaymentRepository
-import com.ass.madhwavahini.domain.utils.StringUtil
+import com.ass.madhwavahini.domain.repository.UserRepository
+import com.ass.madhwavahini.domain.wrapper.StringUtil
 import com.ass.madhwavahini.ui.presentation.payment.PaymentState
 import com.razorpay.PaymentData
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.channels.Channel
-import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.flow.asStateFlow
+import kotlinx.coroutines.delay
 import kotlinx.coroutines.flow.collectLatest
 import kotlinx.coroutines.flow.receiveAsFlow
 import kotlinx.coroutines.launch
@@ -27,29 +29,44 @@ class MainViewModel @Inject constructor(
     private val paymentRepository: PaymentRepository,
     private val userDataStore: UserDataStore,
     private val filesDao: FilesDao,
-    private val application: Application
+    private val application: Application,
+    private val userRepository: UserRepository
 ) : ViewModel() {
 
-    private val _user = MutableStateFlow(User())
-    val user get() = _user.asStateFlow()
+    var user by mutableStateOf(User())
+        private set
 
+    var shouldShowPermissionRational by mutableStateOf(false)
+        private set
 
-    private var _showNotificationRational = Channel<Boolean>()
-    var showNotificationRational = _showNotificationRational.receiveAsFlow()
+    var shouldLogOut by mutableStateOf(false)
+        private set
+
+    fun dismissDialog() {
+        shouldShowPermissionRational = false
+    }
+
+    fun onPermissionResult(
+        isPermissionGranted: Boolean,
+        shouldShowPermissionRationalDialog: Boolean
+    ) {
+        if (!isPermissionGranted && shouldShowPermissionRationalDialog) {
+            shouldShowPermissionRational = true
+        }
+    }
 
     init {
         viewModelScope.launch {
             launch {
                 userDataStore.userData.collectLatest {
-                    _user.value = it
+                    user = it
                 }
             }
         }
     }
 
-    var isLoading = mutableStateOf(false)
+    var isLoading by mutableStateOf(false)
         private set
-
 
     private val _orderState = Channel<PaymentState<Payment>>()
     val orderState get() = _orderState.receiveAsFlow()
@@ -60,7 +77,7 @@ class MainViewModel @Inject constructor(
     fun getOrder() {
         viewModelScope.launch {
             paymentRepository.getPaymentOrder().collectLatest {
-                isLoading.value = it.isLoading
+                isLoading = it.isLoading
                 _orderState.send(it)
             }
         }
@@ -80,15 +97,23 @@ class MainViewModel @Inject constructor(
 
     fun logout() {
         viewModelScope.launch {
-            isLoading.value = true
-            filesDao.getAllFileIds().forEach {
-                val file = File(application.filesDir, "file_$it.txt")
-                if (file.exists()) {
-                    file.delete()
+            userRepository.logoutUser().collectLatest {
+                isLoading = it.isLoading
+                delay(2000)
+                it.data?.let { isLoggedOut ->
+                    if (isLoggedOut) {
+                        filesDao.getAllFileIds().forEach { id ->
+                            val file = File(application.filesDir, "file_$id.txt")
+                            if (file.exists()) {
+                                file.delete()
+                            }
+                        }
+                        userDataStore.logout()
+                    }
+                    shouldLogOut = isLoggedOut
                 }
             }
-            userDataStore.logout()
-            isLoading.value = false
+
         }
     }
 
@@ -99,7 +124,7 @@ class MainViewModel @Inject constructor(
                 paymentId = paymentData.paymentId,
                 paymentSignature = paymentData.signature
             ).collectLatest {
-                isLoading.value = it.isLoading
+                isLoading = it.isLoading
                 _paymentState.send(it)
                 it.data?.let { u ->
                     userDataStore.saveUser(u)
